+++ packages/core/src/interfaces/index.ts
@@ -1,6 +1,7 @@
 export type { LTIClient } from './ltiClient.js';
 export type { LTIConfig } from './ltiConfig.js';
 export type { LTIDeployment } from './ltiDeployment.js';
+export type { LTIDynamicRegistrationSession } from './ltiDynamicRegistrationSession.js';
 export type { LTILaunchConfig } from './ltiLaunchConfig.js';
 export type { LTISession } from './ltiSession.js';
 export type { LTIStorage } from './ltiStorage.js';